Refactor entity objects into dedicated files
